+++ packages/expo/test/index.test.js
@@ -3,14 +3,14 @@ const delivery = require('@bugsnag/delivery-expo')
 jest.mock('expo-constants', () => ({
   default: {
     platform: {},
-    manifest: {}
+    expoConfig: {}
   }
 }))
 
 jest.mock('../../plugin-expo-device/node_modules/expo-constants', () => ({
   default: {
     platform: {},
-    manifest: {}
+    expoConfig: {}
   }
 }))
 
@@ -19,7 +19,7 @@ jest.mock('../../plugin-expo-app/node_modules/expo-application', () => ({}))
 jest.mock('../../plugin-expo-app/node_modules/expo-constants', () => ({
   default: {
     platform: {},
-    manifest: {}
+    expoConfig: {}
   }
 }))